changed the type of variable to be decimal from int

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@ DELIMITER //
 DROP PROCEDURE IF EXISTS ComputeAverageScoreForUser;
 CREATE PROCEDURE ComputeAverageScoreForUser(user_id INT)
 BEGIN
-	DECLARE avg_score INT;
+	DECLARE avg_score DECIMAL;
 	SELECT AVG(score) INTO avg_score FROM corrections WHERE corrections.user_id = user_id;
 	UPDATE users SET average_score = avg_score WHERE id = user_id;
 END;
